+++ dotcom-rendering/src/components/MostViewedRightWithAd.tsx
@@ -1,5 +1,6 @@
 import { css } from '@emotion/react';
 import { ArticleDesign } from '@guardian/libs';
+import { useDeeplyReadTestVariant } from '../lib/useDeeplyReadTestVariant';
 import { RightAdsPlaceholder } from './AdPlaceholder.apps';
 import { AdSlot } from './AdSlot.web';
 import { useConfig } from './ConfigContext';
@@ -18,6 +19,7 @@ type Props = {
  * a right ad slot and a most viewed component
  */
 const MAX_HEIGHT_PX = 1600;
+const MAX_HEIGHT_PX_DEEPLY_READ = 2250;
 
 export const MostViewedRightWithAd = ({
 	format,
@@ -31,6 +33,9 @@ export const MostViewedRightWithAd = ({
 	const useDarkColourScheme =
 		format.design === ArticleDesign.Video ||
 		format.design === ArticleDesign.Audio;
+	const deeplyReadTestVariant = useDeeplyReadTestVariant();
+	const deeplyReadAndMostViewed =
+		deeplyReadTestVariant === 'deeply-read-and-most-viewed';
 
 	return (
 		<div
@@ -42,7 +47,13 @@ export const MostViewedRightWithAd = ({
 				 * On Web - we restrict the height to the maximum height, so that the top right ad can be sticky until the
 				 *          most viewed component is in view at MAX_HEIGHT_PX, or 100% of the article height if it is a short article
 				*/
-				height: ${isApps ? '100%' : `min(100%, ${MAX_HEIGHT_PX}px)`};
+				height: ${isApps
+					? '100%'
+					: `min(100%, ${
+							deeplyReadAndMostViewed
+								? MAX_HEIGHT_PX_DEEPLY_READ
+								: MAX_HEIGHT_PX
+					  }px)`};
 				display: flex;
 				flex-direction: column;
 			`}
